feat(profiles): publish config schema
