+++ go.mod
@@ -4,5 +4,8 @@ go 1.26.1
 
 require (
 	github.com/alchemy/json5 v0.2.0
+	github.com/santhosh-tekuri/jsonschema/v6 v6.0.2
 	gopkg.in/yaml.v3 v3.0.1
 )
+
+require golang.org/x/text v0.14.0 // indirect
